requirements.txt
streamlit==1.36.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
pdfplumber==0.11.4
python-docx==1.1.2
unidecode==1.3.8
